requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
beautifulsoup4==4.12.2
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
lxml==4.9.3
schedule==1.2.0
python-dotenv==1.0.0
selenium==4.15.2
webdriver-manager==4.0.1
